use preavis_id instead of affaire_id + isolate components "comments" & "decision"
