+++ front/src/components/Preavis/preavis.html
@@ -29,11 +29,9 @@
                         <div v-if="item.preavis_attribution">{{ item.preavis_attribution }}</div>
                         <div v-else><md-button class="md-dense md-raised md-primary" @click="preavisAttribution(item)">attribution</md-button></div>
                     </md-table-cell> 
-                    <md-table-cell md-label="Consultation"><md-button class="md-icon-button md-primary" :to="{ name: 'PreavisEdit', params: { id: item.preavis_affaire_id } }"><md-icon>open_in_new</md-icon></md-button></md-table-cell>
+                    <md-table-cell md-label="Consultation"><md-button class="md-icon-button md-primary" :to="{ name: 'PreavisEdit', params: { id: item.preavis_id } }"><md-icon>open_in_new</md-icon></md-button></md-table-cell>
                 </md-table-row>
             </md-table>
         </md-card-content>
     </md-card>
-
 </div>
-
