+++ front/src/components/Preavis/preavisEdit.html
@@ -1,4 +1,4 @@
-<div id="preavisEdit" class="preavisEdit">
+<div id="preavisEdit" class="preavisEdit" v-if="affaireLoaded">
 
     <!-- TITRE -->
     <md-toolbar class="justifyTitleBtn">
@@ -108,74 +108,9 @@
         </md-card-content>
     </md-card>
 
-    <md-card>
-        <md-card-header>
-            <div class="md-title">Conversation</div>
-        </md-card-header>
-        <md-card-content>
-            <div class="md-layout md-gutter md-alignment-top-space-between">
-                <div class="md-layout-item md-size-50">
-                    <md-field>
-                        <label>Commentaire</label>
-                        <md-textarea v-model="commentaire"></md-textarea>
-                    </md-field>
-                    
-                    <div class="md-layout-item" style="float: right;">
-                        <md-button class="md-raised md-primary" @click="saveMessage">Ajouter le commentaire</md-button>
-                    </div>
-                </div>
-                <div class="md-layout-item md-size-50">
-                    <md-table v-model="conversation" md-card>
-                        <md-table-empty-state md-label="Aucun commentaire"></md-table-empty-state>
-                        
-                        <md-table-row slot="md-table-row" slot-scope="{ item }">
-                            <md-table-cell style="width: 15%;" md-label="Date">{{ item.date }}</md-table-cell>
-                            <md-table-cell style="width: 15%;" md-label="Operateur">{{ item.operateur }}</md-table-cell>
-                            <md-table-cell style="width: 70%;" md-label="Message">{{ item.message }}</md-table-cell>
-                        </md-table-row>
-                    </md-table>
-                </div>
-            </div>
-
-        </md-card-content>
-    </md-card>
-
-    <md-card>
-        <md-card-header>
-            <div class="md-title">Décision</div>
-        </md-card-header>
-        <md-card-content>
-            <div style="display: flex;">
-                <div style="width: 200px;">
-                    <div v-if="decisions_liste.length === 0"><em>Aucune décision</em></div>
-                    <div v-else>
-                        <ul>
-                            <li v-for="item in decisions_liste"><a @click="getDecision(item.preavisDecision_id)">{{ item.date }}</a></li>
-                        </ul>
-                    </div>
-                    <md-button class="md-primary md-raised" @click="openNewDecision">Nouveau</md-button>
-                </div>
-                <div style="width: 100%;" v-if="decision.show">
-                <md-card>
-                    <md-card-content>
-                        <div class="md-title" v-if="decision.operateur">Décision du {{ decision.date }} enregistrée par {{ decision.operateur }}</div>
-                        <md-radio v-model="decision.preavis_type_id" :value="1" :disabled="decision.disabled" :class="{highlightDecision: decision.preavis_type_id === 1}">Préavis positif sans condition</md-radio>
-                        <md-radio v-model="decision.preavis_type_id" :value="4" :disabled="decision.disabled" :class="{highlightDecision: decision.preavis_type_id === 4}">Préavis positif avec condition</md-radio>
-                        <md-radio v-model="decision.preavis_type_id" :value="2" :disabled="decision.disabled" :class="{highlightDecision: decision.preavis_type_id === 2}">Préavis négatif</md-radio>
-                        <md-radio v-model="decision.preavis_type_id" :value="3" :disabled="decision.disabled" :class="{highlightDecision: decision.preavis_type_id === 3}">Pas concerné</md-radio>
-                        <md-field>
-                            <label>Remarque</label>
-                            <md-textarea v-model="decision.remarque" :disabled="decision.disabled" style="max-height: 70vh; height: 70vh;"></md-textarea>
-                        </md-field>
-                        <md-button @click="saveDecision" class="md-raised md-primary" style="float: right; margin-bottom: 20px;" :disabled="decision.preavis_type_id == null || decision.remarque == null || decision.remarque === ''" v-if="!decision.operateur">Enregistrer</md-button>
-                    </md-card-content>
-                </md-card>
-                </div>
-            </div>
-
-        </md-card-content>
-    </md-card>
-
+    <preavisEditComments :affaire="affaire" />
+    
+    <preavisEditDecision :affaire="affaire" />
 
     <!-- <div class="rightColumn">
             <md-card class="mapHeight">
