+++ front/src/components/Preavis/preavisEditComments.html
@@ -1,0 +1,35 @@
+<div id="preavisEditComments" class="preavisEditComments">
+
+    <md-card>
+        <md-card-header>
+            <div class="md-title">Conversation</div>
+        </md-card-header>
+        <md-card-content>
+            <div class="md-layout md-gutter md-alignment-top-space-between">
+                <div class="md-layout-item md-size-50">
+                    <md-field>
+                        <label>Commentaire</label>
+                        <md-textarea v-model="commentaire"></md-textarea>
+                    </md-field>
+                    
+                    <div class="md-layout-item" style="float: right;">
+                        <md-button class="md-raised md-primary" @click="saveMessage" :disabled="commentaire === '' || commentaire === null">Ajouter le commentaire</md-button>
+                    </div>
+                </div>
+                <div class="md-layout-item md-size-50">
+                    <md-table v-model="conversation" md-card>
+                        <md-table-empty-state md-label="Aucun commentaire"></md-table-empty-state>
+                        
+                        <md-table-row slot="md-table-row" slot-scope="{ item }">
+                            <md-table-cell style="width: 15%;" md-label="Date">{{ item.date }}</md-table-cell>
+                            <md-table-cell style="width: 15%;" md-label="Operateur">{{ item.operateur }}</md-table-cell>
+                            <md-table-cell style="width: 70%;" md-label="Message">{{ item.message }}</md-table-cell>
+                        </md-table-row>
+                    </md-table>
+                </div>
+            </div>
+
+        </md-card-content>
+    </md-card>
+
+</div>
