+++ front/src/components/Preavis/preavisEditDecision.css
@@ -1,0 +1,3 @@
+.highlightDecision {
+    background-color: yellow;
+}
