+++ front/src/components/Preavis/preavisEditDecision.html
@@ -1,0 +1,37 @@
+<div id="preavisEditDecision" class="preavisEditDecision">
+    <md-card>
+        <md-card-header>
+            <div class="md-title">Décision</div>
+        </md-card-header>
+        <md-card-content>
+            <div style="display: flex;">
+                <div style="width: 200px;">
+                    <div v-if="decisions_liste.length === 0"><em>Aucune décision</em></div>
+                    <div v-else>
+                        <ul>
+                            <li v-for="item in decisions_liste"><a @click="getDecision(item.preavisDecision_id)">{{ item.date }}</a></li>
+                        </ul>
+                    </div>
+                    <md-button class="md-primary md-raised" @click="openNewDecision">Nouveau</md-button>
+                </div>
+                <div style="width: 100%;" v-if="decision.show">
+                <md-card>
+                    <md-card-content>
+                        <div class="md-title" v-if="decision.operateur">Décision du {{ decision.date }} enregistrée par {{ decision.operateur }}</div>
+                        <md-radio v-model="decision.preavis_type_id" :value="1" :disabled="decision.disabled" :class="{highlightDecision: decision.preavis_type_id === 1}">Préavis positif sans condition</md-radio>
+                        <md-radio v-model="decision.preavis_type_id" :value="4" :disabled="decision.disabled" :class="{highlightDecision: decision.preavis_type_id === 4}">Préavis positif avec condition</md-radio>
+                        <md-radio v-model="decision.preavis_type_id" :value="2" :disabled="decision.disabled" :class="{highlightDecision: decision.preavis_type_id === 2}">Préavis négatif</md-radio>
+                        <md-radio v-model="decision.preavis_type_id" :value="3" :disabled="decision.disabled" :class="{highlightDecision: decision.preavis_type_id === 3}">Pas concerné</md-radio>
+                        <md-field>
+                            <label>Remarque</label>
+                            <md-textarea v-model="decision.remarque" :disabled="decision.disabled" style="max-height: 70vh; height: 70vh;"></md-textarea>
+                        </md-field>
+                        <md-button @click="saveDecision" class="md-raised md-primary" style="float: right; margin-bottom: 20px;" :disabled="decision.preavis_type_id == null || decision.remarque == null || decision.remarque === ''" v-if="!decision.operateur">Enregistrer</md-button>
+                    </md-card-content>
+                </md-card>
+                </div>
+            </div>
+
+        </md-card-content>
+    </md-card>
+</div>
